refactor(prompt-area): extract list-ops and paste-resolution from engine
Splits prompt-area-engine.ts (29KB) into three cohesive modules:
- prompt-area-engine.ts (~18KB) — core segment ops, chip ops, text replace,
  markdown, segmentsEqual. mergeAdjacentTextSegments now exported for the
  two new sibling modules.
- prompt-area-list-ops.ts (new) — getListContext + 6 list handlers +
  normalizeListPrefixes + ListContext type.
- prompt-area-paste-resolution.ts (new) — resolveTriggersInSegments and its
  private splitTextByTriggerPatterns helper.

Updates registry.json with 2 new entries (registry:component, matching
every other block file). Bumps registry.test.ts file-count from 15 to 17.

All 593 tests pass; behavior identical.
