+++ registry/new-york/blocks/prompt-area/__tests__/prompt-area-engine.test.ts
@@ -7,20 +7,22 @@ import {
   resolveChip,
   removeChipAtIndex,
   revertChipAtIndex,
-  resolveTriggersInSegments,
   replaceTextRange,
+  isValidTriggerPosition,
+  parseInlineMarkdown,
+  toggleMarkdownWrap,
+  segmentsEqual,
+} from '../prompt-area-engine'
+import {
   getListContext,
   autoFormatListPrefix,
   insertListContinuation,
   indentListItem,
   outdentListItem,
   removeListPrefix,
-  isValidTriggerPosition,
-  parseInlineMarkdown,
-  toggleMarkdownWrap,
-  segmentsEqual,
   normalizeListPrefixes,
-} from '../prompt-area-engine'
+} from '../prompt-area-list-ops'
+import { resolveTriggersInSegments } from '../prompt-area-paste-resolution'
 
 // ---------------------------------------------------------------------------
 // Test trigger configs
